+++ src/main/kotlin/lotto/model/game/LottoNumber.kt
@@ -1,4 +1,4 @@
-package lotto.model
+package lotto.model.game
 
 data class LottoNumber(val number: Int) {
     init {
