+++ src/main/kotlin/lotto/model/game/WinningLotto.kt
@@ -1,3 +1,3 @@
-package lotto.model
+package lotto.model.game
 
 class WinningLotto(val winningLotto: Lotto, val bonusNumber: LottoNumber)
