+++ src/main/kotlin/lotto/model/result/Result.kt
@@ -1,3 +1,3 @@
-package lotto.model
+package lotto.model.result
 
 class Result(val coincidence: Coincidence, val matchCount: Int)
