+++ src/test/kotlin/lotto/model/CoincidenceTest.kt
@@ -1,5 +1,10 @@
 package lotto.model
 
+import lotto.model.game.Lotto
+import lotto.model.game.LottoNumber
+import lotto.model.game.Lottos
+import lotto.model.game.WinningLotto
+import lotto.model.result.Coincidence
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
