+++ src/test/kotlin/lotto/model/DelimiterTest.kt
@@ -1,5 +1,6 @@
 package lotto.model
 
+import lotto.model.input.Delimiter
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
